findByID의 입력 인자 데이터 타입 변경

+++ jjikeobang-server/src/main/java/com/jjikeobang/member/repository/MemberRepository.java
@@ -15,7 +15,7 @@ public interface MemberRepository {
 
     void putMember(JoinMemberDTO member);
 
-    Member findById(int memberId);
+    Member findById(long memberId);
 
     Member findByLoginId(String userId);
 }
